AI_Registration_Assistant/backend/requirements.txt
# Core
fastapi>=0.68.0,<0.69.0
uvicorn[standard]>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.0.7
python-dotenv>=1.0.0,<2.0.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.7.0,<2.0.0
pydantic>=1.8.0,<2.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
gunicorn>=20.1.0,<21.0.0
email-validator>=1.1.3,<2.0.0
python-slugify>=5.0.2,<6.0.0
python-magic>=0.4.24,<0.5.0
python-magic-bin>=0.4.14,<0.5.0; sys_platform == 'win32'  # Windows support
httpx>=0.18.2,<0.19.0  # For async HTTP requests if needed
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email
python-email-validator==2.0.0

# Production
gunicorn==21.2.0

# Development (not installed in production)
pytest==7.4.3
httpx==0.25.1
